+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.46.0 - 2026-01-20
+
+### Enhancements
+- Added new off-market publisher for Cboe Futures Exchange (`XCBF_PITCH_XOFF`)
+- Added new `StatType` variants to be used by `XCBF.PITCH` dataset:
+  - `UpperPriceLimit`
+  - `LowerPriceLimit`
+  - `BlockVolume`
+  - `VenueSpecificVolume1`
+
 ## 0.45.0 - 2025-12-10
 
 ### Enhancements